Convert schematic to KiCad 5
In KiCad 5, symbols reference the library they are in explicitly. Also,
a lot of symbols were changed for more consistency. This commit changes
all symbols to their new version in the official libraries, removing the
need for our own symbol library (which is removed by this commit).

For the STU MOSFET component, not specific symbol is available, so a
generic MOSFET symbol is used instead.

This also updates the footprint fields to point to versions from the
KiCad 5 libraries, but does not update the actual layout to use them
yet.

For the GPS connector symbol, the pinout was reversed (matching the,
technically incorrect, footprint used). This fixes the pin order in the
schematic and updates the footprint field to a footprint that matches
this pin order, so when a netlist is imported in the future, this should
still match.