Add DC plug and Lipo connector to schematic
These were in the board layout, but not in the schematic. This adds them
to the schematic, and sets the appropriate references in the layouted
footprints.

Note that the DC plug symbol has a different pin numbering than the
footprint currently in use, but the symbol does match the footprint set
in the schematic, so this should work out when the netlist is re-read in
the future.