Convert board layout to KiCad 5
This replaces all footprints with new versions from KiCad 5. Some
positions and labeling as changed, so this updates everything to match
the original layout for the most part. Some footprints (barrel jack,
regulator, battery holder, solder jumpers, perfboard "connector") are
slightly different. In particualr the battery holde slightly changes the
mounting hole position and size, so this might need to be changed back
later.

The only footprints that are still used from the project library are the
logo, all others are removed. All components are also in the schematic,
except for the perfboard individiual pins (which also do not have any
external footprint they refer to).

The GPS connector footprint has holes that are slightly (0.05mm) bigger,
which should make it easier to fit the connector.